Move workflow runner to test support
